+++ Changelog.md
@@ -2,6 +2,7 @@ AppleALC Changelog
 ==================
 #### v1.7.0
 - Fix headphones after sleep on Latitude 7390 2-in-1 (ALC225 layout 30)
+- Added `dump_coeff.sh` script in `Tools` to dump processing caps under macOS, plus docs in Wiki
 
 #### v1.6.9
 - Added 0x100003 revision for ALCS1220A